Update OneButton.cpp
For the single-click event, I think it's more reasonable to confirm the event if there is no further click within 200 milliseconds after the button is released. This way, the single-click does not need to wait too long, otherwise, it would have to wait until the _click_ms threshold to confirm the click.

For the long-press event, when the long-press event is detected, record the current time. After the button is released, compare the time. If the release happens within the _click_ms window, return to the "released" state. The benefit of this is that the long-press time can be set shorter without being constrained by _click_ms. Otherwise, according to the original logic, _press_ms would almost never be shorter than _click_ms. The benefit is that even after entering the long-press state, we can still continue to check for single-click, double-click, or multi-click within the _click_ms window.